Renamed and moved email Freemarker classes to avoid confusion with PDf Freemarker functions
